requirements.txt
Keras_Preprocessing==1.1.2
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
streamlit==1.24.0
torch==2.0.1
transformers==4.30.2
underthesea==6.2.0
